Fixes & updates for installing JDK 8_60 on SUSE 13.2.
In keeping with the spirit of SUSE, removed use of sudo; run as root.

Replaced variable ORACLE_JAVA (/usr/java), which 13.2 seems not to use, with JVM_FOLDER (/usr/lib(64)/jvm). This makes the "main link" between these two superfluous, so it is gone.

Added an alternative update for JDK API documentation, <javadocdir>.

Adjusted script comments to reflect these changes. Included a patch for existing installs.